refactor: 북마크 href/공유·정렬 헬퍼를 bookmark-core.js로 분리 (ADR-034 후속 PR2)
bookmark.js의 BOOKMARK_HREF + BOOKMARK_SORT 블록(href/share 빌더 + per-device
정렬·최근열람 헬퍼, localStorage 기반)을 js/app/bookmark-core.js(leaf)로 분리.
bookmark.js 3,364→3,235줄.

- 전부 bookmark 내부 전용(외부 호출 0, window facade·appBookmark·export 등장 0)
  이라 깔끔한 import-only 이동. bookmark이 쓰는 7개(_bookmarkHref·_buildSharePayload·
  getBookmarkSort·setBookmarkSort·markBookmarkViewed·_forgetViewed·sortBookmarkNodes)
  만 export, 내부 헬퍼(_loadViewedMap·_nodeTitle·_bookmarkComparator·SITE_BASE)는
  core 내부 유지.
- bookmark.test.js의 BOOKMARK_HREF/SORT 슬라이스를 bookmark-core.js로(ACTIVE 로더는
  core의 HREF + bookmark의 ACTIVE 연결). index.html·sw.js 매니페스트 추가.

검증: tsc 0, 유닛 678, e2e bookmark 17, playwright 로드 pageerror 0(지난 export
버그 재발 방지 검사).
