+++ tests/unit/bookmark.test.js
@@ -39,6 +39,10 @@ const BOOKMARK_SOURCE = fs.readFileSync(BOOKMARK_PATH, "utf8");
 // marker block now lives there, so its loader slices from this source.
 const VERSE_SPEC_PATH = path.resolve(__dirname, "../../js/app/verse-spec.js");
 const VERSE_SPEC_SOURCE = fs.readFileSync(VERSE_SPEC_PATH, "utf8");
+// Href/share + sort helpers moved to bookmark-core.js (ADR-034 후속 PR2); their
+// marker blocks (BOOKMARK_HREF / BOOKMARK_SORT) now live there.
+const BOOKMARK_CORE_PATH = path.resolve(__dirname, "../../js/app/bookmark-core.js");
+const BOOKMARK_CORE_SOURCE = fs.readFileSync(BOOKMARK_CORE_PATH, "utf8");
 
 function extractBlock(name, source = BOOKMARK_SOURCE) {
   const begin = `// ── BEGIN ${name} ──`;
@@ -1117,7 +1121,7 @@ test('closeSwipedRowIfOutside: non-Node target → guard rejects, close fires', 
 function loadBookmarkHref() {
   const ctx = { Object, Array, String, console, Error };
   vm.createContext(ctx);
-  vm.runInContext(extractBlock("BOOKMARK_HREF"), ctx, { filename: "bookmark-href.js" });
+  vm.runInContext(extractBlock("BOOKMARK_HREF", BOOKMARK_CORE_SOURCE), ctx, { filename: "bookmark-core.js" });
   return { _bookmarkHref: ctx._bookmarkHref, _buildSharePayload: ctx._buildSharePayload };
 }
 
@@ -1128,7 +1132,7 @@ function loadBookmarkHref() {
 function loadBookmarkActive() {
   const ctx = { Object, Array, String, console, Error };
   vm.createContext(ctx);
-  vm.runInContext(extractBlock("BOOKMARK_HREF") + "\n" + extractBlock("BOOKMARK_ACTIVE"), ctx, {
+  vm.runInContext(extractBlock("BOOKMARK_HREF", BOOKMARK_CORE_SOURCE) + "\n" + extractBlock("BOOKMARK_ACTIVE"), ctx, {
     filename: "bookmark-active.js",
   });
   return {
@@ -1564,7 +1568,7 @@ function loadBookmarkSort(seed = {}) {
     localStorage,
   };
   vm.createContext(ctx);
-  vm.runInContext(extractBlock("BOOKMARK_SORT"), ctx, { filename: "bookmark-sort.js" });
+  vm.runInContext(extractBlock("BOOKMARK_SORT", BOOKMARK_CORE_SOURCE), ctx, { filename: "bookmark-core.js" });
   return {
     getBookmarkSort: ctx.getBookmarkSort,
     setBookmarkSort: ctx.setBookmarkSort,
